reorganise meson, file names, and fully split kernel-fuse source
