+++ src/uenv/config.h.in
@@ -1,11 +1,13 @@
 #pragma once
 
+// clang-format off
 #define UENV_VERSION "@version@"
 #define UENV_VERSION_MAJOR @version_major@
 #define UENV_VERSION_MINOR @version_minor@
 #define UENV_VERSION_PATCH @version_patch@
 #define UENV_VERSION_PRERELEASE "@version_prerelease@"
 
-// 1 when built with the FUSE/squashfuse backend, 0 for the setuid/kernel
+// 1 when built with the FUSE/squashfuse backend, 0 for the kernel/setuid
 // backend. Selected by the meson `backend` option.
 #define UENV_BACKEND_FUSE @backend_fuse@
+// clang-format on
